server/requirements.txt
alembic==1.7.5
anyio==3.4.0
asgiref==3.4.1
bcrypt==3.2.0
black==21.12b0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cryptography==36.0.1
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.70.1
fastapi-camelcase==1.0.5
flake8==4.0.1
google-auth==2.3.3
greenlet==1.1.2
h11==0.12.0
httpcore==0.14.3
httptools==0.3.0
idna==3.3
isort==5.10.1
itsdangerous==2.0.1
Mako==1.1.6
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.4.0
psycopg2==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.8.2
pyflakes==2.4.0
pyhumps==3.0.2
python-dotenv==0.19.2
python-http-client==3.3.4
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.26.0
rfc3986==1.5.0
rsa==4.8
sendgrid==6.9.3
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.28
starkbank-ecdsa==2.0.3
starlette==0.16.0
tomli==1.2.3
typing_extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
